# docs/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: Machine learning system for detecting anomalies in children's drawings
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Children's Drawing Anomaly Detection System
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/alerts/thresholds:
    post:
      description: Update system alert thresholds.
      operationId: update_alert_thresholds_api_v1_alerts_thresholds_post
      requestBody:
        content:
          application/json:
            example: {}
            schema:
              additionalProperties:
                type: number
              title: Thresholds
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update alert thresholds
      tags:
      - health
...
servers:
- description: Development server
  url: http://localhost:8000
- description: API v1 base path
  url: /api/v1